requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT[crypto]==2.8.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
httpx==0.24.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Supabase integration
supabase==2.0.2
postgrest==0.13.0

# AWS services
boto3==1.34.0
botocore==1.34.0
aioboto3==12.3.0

# Akeyless integration
akeyless==3.0.0

# Additional utilities
cryptography==41.0.7
python-jose[cryptography]==3.3.0
email-validator==2.1.0
pytz==2023.3 

# AI Analysis Dependencies
openai>=1.0.0

# PDF and Image Processing Dependencies
pdfplumber==0.10.3
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0